requirements.txt
easyocr
streamlit
transformers
pillow
torch
accelerate